Treat a null elementNs in getPropertyType/geAtributeType as the HTML namespace
This is the behavior that Chromium and WebKit currently implement, see
https://phabricator.services.mozilla.com/D226547

Currently, it seems that the "interface for localName and elementNs"
would fallback to `Element`, similarly to what happens with
`document.createElementNS(null, "div")` for example.

Note that a null attrNs is already handled by the spec.

See also https://github.com/w3c/trusted-types/issues/553 for details.