Put trusted type first when declaring union in IDL files
This is for consistency with how it is done in the HTML specification
and in the PR to DOM (https://github.com/whatwg/dom/pull/1268). This
is not web-exposed but may matter for implementations. For example,
they may generate different C++ structure to represent
`(TrustedHTML or DOMString)` and `(DOMString or TrustedHTML)`.